Add environment variables to Helm chart + Move handlers into different package
