+++ chart/templates/deployment.yaml
@@ -4,7 +4,6 @@ metadata:
   name: {{ include "keptn-service.fullname" . }}
   labels:
     {{- include "keptn-service.labels" . | nindent 4 }}
-
 spec:
   replicas: 1
   selector:
@@ -47,6 +46,25 @@ spec:
             value: "http://configuration-service:8080"
           - name: env
             value: 'production'
+          - name: K8S_POD_NAME
+            valueFrom:
+              fieldRef:
+                apiVersion: v1
+                fieldPath: metadata.name
+          - name: K8S_NAMESPACE
+            valueFrom:
+              fieldRef:
+                apiVersion: v1
+                fieldPath: metadata.namespace
+          - name: K8S_NODE_NAME
+            valueFrom:
+              fieldRef:
+                apiVersion: v1
+                fieldPath: spec.nodeName
+          - name: POD_NAMESPACE
+            valueFrom:
+              fieldRef:
+                fieldPath: metadata.namespace
           {{- if .Values.remoteControlPlane.enabled }}
           - name: KEPTN_API_ENDPOINT
             value: "{{ .Values.remoteControlPlane.api.protocol }}://{{ .Values.remoteControlPlane.api.hostname }}/api"
