+++ handler/action_triggered_event_handler.go
@@ -1,0 +1,60 @@
+package handler
+
+import (
+	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
+	"github.com/keptn/go-utils/pkg/sdk"
+	"time"
+)
+
+type actionTriggeredEventHandler struct {
+}
+
+func NewActionTriggeredEventHandler() *actionTriggeredEventHandler {
+	return &actionTriggeredEventHandler{}
+}
+
+// Execute handles action.triggered events
+// TODO: Add in your handler code
+func (g *actionTriggeredEventHandler) Execute(k sdk.IKeptn, event sdk.KeptnEvent) (interface{}, *sdk.Error) {
+	actionTriggeredEvent := &keptnv2.ActionTriggeredEventData{}
+
+	if err := keptnv2.Decode(event.Data, actionTriggeredEvent); err != nil {
+		return nil, &sdk.Error{Err: err, StatusType: keptnv2.StatusErrored, ResultType: keptnv2.ResultFailed, Message: "failed to decode action.triggered event: " + err.Error()}
+	}
+
+	k.Logger().Infof("Handling Action Triggered Event: %s", event.ID)
+	k.Logger().Infof("Action=%s", actionTriggeredEvent.Action.Action)
+
+	// check if action is supported
+	if actionTriggeredEvent.Action.Action == "action-xyz" {
+		k.Logger().Info("Action remediation triggered")
+		// -----------------------------------------------------
+		// TODO: Implement your remediation action here
+		// -----------------------------------------------------
+		time.Sleep(1 * time.Second) // Example: Wait 5 seconds. Maybe the problem fixes itself.
+
+		// Return finished event
+		finishedEventData := getActionFinishedEvent(keptnv2.ResultPass, keptnv2.StatusSucceeded, *actionTriggeredEvent, "")
+		k.Logger().Infof("Finished event: %o", finishedEventData)
+
+		return finishedEventData, nil
+	} else {
+		k.Logger().Infof("Retrieved unknown action %s, skipping...", actionTriggeredEvent.Action.Action)
+		return nil, nil
+	}
+}
+
+func getActionFinishedEvent(result keptnv2.ResultType, status keptnv2.StatusType, actionTriggeredEvent keptnv2.ActionTriggeredEventData, message string) keptnv2.ActionFinishedEventData {
+
+	return keptnv2.ActionFinishedEventData{
+		EventData: keptnv2.EventData{
+			Project: actionTriggeredEvent.Project,
+			Stage:   actionTriggeredEvent.Stage,
+			Service: actionTriggeredEvent.Service,
+			Labels:  actionTriggeredEvent.Labels,
+			Status:  status,
+			Result:  result,
+			Message: message,
+		},
+	}
+}
